make in-process program's output deterministic
When an in-process program writes to both stdout and stderr, one at a
time, sequentially, the output is non-deterministic. More specifically,
the order of writing might not be respected.

This happens because there is one pipe for each file, and each reading
pipe is fetched concurrently (each in its own goroutine), in order to be
subsequently copied to a locking writer.

Since there is a goroutine for each reading pipe to be copied to the
locking writer, a pipe that was written later might end up copied before
a pipe that was written earlier.

The locking writer does prevent race condition between writes but
doesn't solve the order problem. The solution, then, is to have a single
pipe for both files, allowing sequential writing to be guaranteed while
also unifying both outputs deterministically in a single buffer.