swap diff outputs
The current output shows want=+ and got=-, which is really confusing,
since minus denotes elements missing and plus denotes additional
elements, according to go-cmp documentation:

  https://pkg.go.dev/github.com/google/go-cmp/cmp?tab=doc#Diff

Also, in that same documentation, the example provided for Diff uses
minus (-) for "want" and plus (+) for "got". It is also used this way in
other languages.

For comparing values, the "got/want" order makes sense (e.g. errors.Is),
but for printing diffs, using "want/got" makes more sense and is easier
to read.