support checking exit codes
Allow tests to specify exit codes on failure, and check
them against the actual exit code.

For built-ins this will just work, because we look for the wrapped
sycall.Errno. For InProcessProgram and user-defined CommandFuncs,
define an error type that holds the exit code and allows it to
be retrieved.